Middleware/requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2
et-xmlfile==1.1.0
gevent==20.9.0
greenlet==0.4.17
idna==2.10
iniconfig==1.1.1
numpy==1.21.0
openpyxl==3.0.7
packaging==20.4
pandas==1.1.5
Pillow==8.1.2
pip==21.1.3
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.24.0
setuptools==50.3.2
six==1.15.0
toml==0.10.1
urllib3==1.25.11
websocket==0.2.1
websocket-client==0.57.0
XlsxWriter==3.0.1
zope.event==4.5.0
zope.interface==5.1.2
